% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-log_into.R
\name{dynamic_url}
\alias{dynamic_url}
\title{Dynamic url function}
\usage{
dynamic_url(doctor = "endokrynolog", region = "warszawa")
}
\arguments{
\item{doctor}{Type polish name of doctor you are looking for}

\item{region}{type city in which you want to find Your doctor}
}
\description{
This function allows you to create search url based on condition you add in arguments
}
\examples{
dynamic_url(doctor = "endokrynolog")
}
