% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-scrap_medicover_data.R
\name{get_content}
\alias{get_content}
\title{get_content function}
\usage{
get_content(content = box_content)
}
\arguments{
\item{content}{Elemnt containing box with doctor data}
}
\description{
This function get doctor name and location
}
\examples{
get_content()
}
