% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-scrap_medicover_data.R
\name{get_head}
\alias{get_head}
\title{get_head function}
\usage{
get_head(box = box_head)
}
\arguments{
\item{box}{Elemnt containing head with time data}
}
\description{
This function get doctors time availability data
}
\examples{
get_head()
}
