% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-log_into.R
\name{get_ids_list}
\alias{get_ids_list}
\title{get_ids_list url function}
\usage{
get_ids_list(field_name = "SpecializationId")
}
\arguments{
\item{field_name}{name of field that you want to listr}
}
\description{
This function allows you to list field values with its names
}
\examples{
get_ids_list(doctor = "endokrynolog")
}
