% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-scrap_medicover_data.R
\name{get_reservation_data}
\alias{get_reservation_data}
\title{get_reservation_data function}
\usage{
get_reservation_data(sleep_time = 8, click_break = 2)
}
\arguments{
\item{sleep_time}{number of seconds to wait untill medicover website get more doctors details}

\item{click_break}{number of seconds to wait untill more elemnt is getting clicked}
}
\description{
This function scrap doctor availiability data from medicover service
}
\examples{
get_reservation_data(sleep_time = 8, click_break = 2 )
}
