% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-log_into.R
\name{log_into_medicover}
\alias{log_into_medicover}
\title{log_into_medicover}
\usage{

  log_into_medicover(url = "https://mol.medicover.pl/MyVisits?regionId=204&bookingTypeId=2&specializationId=178&serviceId=&clinicId=-1&languageId=-1&doctorId=-1&searchSince=2017-07-14T08\%3A00\%3A00.000Z&periodOfTheDay=0&isSetBecauseOfPcc=false&isSetBecausePromoteSpecialization=false&selectedSpecialties=",
  credentials_path = "~/pass/medicover.txt",
  submit_id = "//*[@id=\\"oidc-submit\\"]")
}
\arguments{
\item{url}{Type polish name of doctor you are looking for}

\item{credentials_path}{path to file with user login and pass}

\item{submit_id}{id value for submit button in medicover dom}
}
\description{
This function allows you to log into medicover service
}
\examples{
log_into_medicover(credentials_path = "~/pass/medicover.txt")
}
