% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-aditional_functions.R
\name{read_credntials}
\alias{read_credntials}
\title{Read_credentials function}
\usage{
read_credntials(credentials_path = "~/pass/medicover.txt")
}
\arguments{
\item{credentials_path}{Address to txt file with login and password}
}
\description{
This function read credentials from txt file
}
\examples{
read_credntials()
}
