% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-get_doctor_grade.R
\name{scrap_doctor_rate}
\alias{scrap_doctor_rate}
\title{scrap_doctor_rate function}
\usage{
scrap_doctor_rate(doctor_name)
}
\arguments{
\item{doctor_name}{Doctor name and surname scrapped from medicover website}
}
\description{
This function allows you to scrap doctors grade from znanylekarz based on given names
}
\examples{
scrap_doctor_rate(doctor_name = "Kowalski Jan")
}
